bug fixed

+++ remove.sh
@@ -1,16 +1,16 @@
 #!/bin/bash
-if [ $# != 1 ]
+if [ $# -ne 1 ]
 	then 
 		echo "utilizzo errato: remove <file_name>"
 		exit;
 fi
 
 var=$(find . -name $@)
-arrIN=(${var// / })
+arrIN=(${var// / }) #versione decente da cercare o che minchia vuol dire
 
 n=${#arrIN[@]}
 
-if [ '$n' = '2' ]		# caso singolo file trovato
+if [ $n -eq 2 ]		# caso singolo file trovato
 	then
 		rm -f $var;
 else
@@ -25,6 +25,11 @@ else
 	echo "A causa di multiplo ritrovamento Inserisci nome repo"
 	
 	read name
+	if [ ! -d ./$name ]
+		then
+			echo  "<repo> "$name" non esistente "
+			exit;
+	fi
 
-	rm -f $(echo -e $var | grep -w "\<$name\>")
+	rm -f $(echo -e $var | grep -w "\<$name\>") # da controllare la rindondanza di -w e \< \>
 fi
